refactor: add interface for payload message collator
