+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/injection/SessionModule.kt
@@ -7,6 +7,7 @@ import io.embrace.android.embracesdk.session.caching.PeriodicSessionCacher
 import io.embrace.android.embracesdk.session.message.PayloadFactory
 import io.embrace.android.embracesdk.session.message.PayloadFactoryImpl
 import io.embrace.android.embracesdk.session.message.PayloadMessageCollator
+import io.embrace.android.embracesdk.session.message.PayloadMessageCollatorImpl
 import io.embrace.android.embracesdk.session.orchestrator.OrchestratorBoundaryDelegate
 import io.embrace.android.embracesdk.session.orchestrator.SessionOrchestrator
 import io.embrace.android.embracesdk.session.orchestrator.SessionOrchestratorImpl
@@ -43,7 +44,7 @@ internal class SessionModuleImpl(
 ) : SessionModule {
 
     override val payloadMessageCollator: PayloadMessageCollator by singleton {
-        PayloadMessageCollator(
+        PayloadMessageCollatorImpl(
             essentialServiceModule.configService,
             essentialServiceModule.metadataService,
             dataContainerModule.eventService,
